Update README.md
I tried to redeploy all parts of code that github failed to deploy before. if the files i provided are downloaded and then structured as outlined in directory structure you should be able to view an inventory list as visitor, and then if you sign in or log in to a manager account you should be able to add, delete, or edit items within that inventory list. Items and created accounts are written in the data base as a many to one relationship, and you should only be able to log in to accounts that have earlier been created. 

if GitHub again fails, i have hosted just the HTML files with a different website builder, you can view the UI there. However, since this other site only stores HTML's JS's and CSS's, whenever the user clicks a button that would normally use java to interact with the data base, i have the HTML just send an alert saying what the JS would be doing to the DB. 

https://1922167.playcode.io/